Add gaussian likelihood function to likelihoods
This adds the `build_gaussian` function to forecasting.likelihoods,
which is an implementation of the gaussian likelihood by accepting
the data, model parameters (in this case this is just the mean) and
the covariance matrix. All are expected to be numpy arrays.

The data is expected to be a 2D array, which corresponds to an array
with arrays of data points (each array being the values of the C_ls,
for example). In the case of a 1D forecast the function casts the data
to a 2D array, which should simplify use a bit.

It turns out to be necessary that, to compute the actual likelihood,
the data have to be combined into a box of coordinates. I have opted
to do this purely in the `build_gaussian` function as it is currently
the only function that makes use of it. However, I will probably
abstract that to a separate function later, as I expect the other
likelihood functions will have need of it too.

Another thing to note is that I have chosen to do some verification
of the inputs in the function, which takes the form of a number of
if statements. Probably there is a better way to do this, or perhaps
it is sufficient to put these statements in a function if it turns
out that the other likelihood functions need them. For now I think
this is sufficient.