Pass command line arguments to pytest (#273)
This change passes allows command line arguments for pytest to be passed
through tox.

Whilst tox is used as a workflow runner, we use pytest to run the actual
tests. Since we have a lot of tests, it is useful to be able to control
pytest directly. For example, to run a specific test instead of the
entire test suite.