Add logger (#311)
This PR replaces the warnings and print statements with calls to a
logger.

The setup is a bit limited, as we currently don't have many of such log
messages. We may want to add more in the future, though, so it is good
to have something like this.

I feel it makes sense to switch to this logging module since derivkit is
ultimately a library. The `logging` module allows users to configure the
way the logs are displayed. In the interest of keeping this small I
haven't added facilities in derivkit itself. It seems to be recommended
for libraries to not configure their own logging, but it may be useful
to add a few convenience functions for users at some point if there is a
demand for them.

I have added some documentation about the derivkit logger and how to
configure it.