Resolve docstring indentation issue and configure Sphinx-multiversion (#316)
This PR fixes an indentation issue in one of the docstrings. It also
adds rules for which docs have to be built. With this change
sphinx-multiversion will only build the documentation for the main
branch and the version tags (tags of the form `vX.Y.Z`, with `X`, `Y`,
`Z` an integer.